requirements.txt
python-telegram-bot
yt-dlp
browser-cookie3
